Eliminate duplicate PrimeTeleOp by moving logic to shared base class
Both competition TeleOps were byte-for-byte identical except for package
and @TeleOp name. Extracted all logic into PrimeTeleOpBase (src/main),
added RobotConfig interface for per-robot tuning values, and reduced each
flavor's PrimeTeleOp to a 10-line subclass that names the team and
provides its config.
