+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/config/RobotConfig.java
@@ -1,0 +1,7 @@
+package org.firstinspires.ftc.teamcode.common.config;
+
+public interface RobotConfig {
+    double defaultBaseSpeed();
+    double defaultSensitivity();
+    double defaultDeadzone();
+}
