+++ TeamCode/src/team22091/java/org/firstinspires/ftc/teamcode/team22091/Team22091Config.java
@@ -1,0 +1,9 @@
+package org.firstinspires.ftc.teamcode.team22091;
+
+import org.firstinspires.ftc.teamcode.common.config.RobotConfig;
+
+public class Team22091Config implements RobotConfig {
+    @Override public double defaultBaseSpeed()   { return 0.85; }
+    @Override public double defaultSensitivity() { return 1.2;  }
+    @Override public double defaultDeadzone()    { return 0.1;  }
+}
